+++ R/map.R
@@ -1,7 +1,7 @@
 #' Create a leaflet map.
 #'
 #' @usage map_leaflet(data, color = "#ff3399",
-#'   provider_tiles = "OpenStreetMap.BlackAndWhite", popup = function(x) { x["id"] },
+#'   provider_tiles = "Esri.WorldGrayCanvas", popup = function(x) { x["id"] },
 #'   antarctic = FALSE)
 #' @param data the occurrences from \code{occurrence()}.
 #' @param color color to be used for the dots.
@@ -9,7 +9,7 @@
 #' @param popup function generating the popup content.
 #' @param antarctic use antarctic polar stereographic projection.
 #' @export
-map_leaflet <- function(data, color = "#ff3399", provider_tiles = "OpenStreetMap.BlackAndWhite",
+map_leaflet <- function(data, color = "#ff3399", provider_tiles = "Esri.WorldGrayCanvas",
                         popup = function(x) { x["id"] }, antarctic = FALSE) {
   p <- NULL
   if (!is.null(popup)) {
@@ -63,4 +63,18 @@ map_ggplot <- function(data, color = "#ff3399") {
     xlab("longitude") +
     ylab("latitude") +
     coord_quickmap()
+  m
+}
+
+#' Get a WKT geometry by drawing on a map.
+#'
+#' @usage get_geometry(provider_tiles = "Esri.WorldGrayCanvas")
+#' @param provider_tiles the base map provider.
+#' @export
+get_geometry <- function(provider_tiles = "Esri.WorldGrayCanvas") {
+  s <- editMap(leaflet() %>% addProviderTiles(provider_tiles), editorOptions = list(polylineOptions = FALSE, markerOptions = FALSE, circleOptions = FALSE, circleMarkerOptions = FALSE), title = "Draw geometry")$all
+  if (!is.null(s)) {
+    return(st_as_text(st_combine(s$geometry)))
+  }
+  return(NULL)
 }